chore: migrate eslint to v9 flat config
- bump eslint 8.57 -> 9.39, add @eslint/js, globals, typescript-eslint, angular-eslint
- replace abandoned eslint-plugin-rxjs with eslint-plugin-rxjs-x
- replace abandoned eslint-plugin-eslint-comments with @eslint-community fork
- drop direct @typescript-eslint/* and @angular-eslint/* deps (provided by umbrella packages)
- convert .eslintrc.js + .eslintignore into eslint.config.js
- drop the removed --ext flag, add --no-warn-ignored to lint-staged
