--- .eslintignore
@@ -1,11 +1,0 @@
-dist
-node_modules
-tmp
-.angular
-playwright-report
-
-# eslint ignores root level "dot" files by default
-!/.*.js
-
-# ignore index.html
-**/index.html
