+++ .lintstagedrc.js
@@ -1,5 +1,5 @@
 module.exports = {
     '*': 'prettier --write --ignore-unknown',
-    '*.{js,ts,html}': 'eslint --max-warnings=0 --fix',
+    '*.{js,ts,html}': 'eslint --max-warnings=0 --no-warn-ignored --fix',
     '*.{css,scss}': 'stylelint --max-warnings=0 --fix'
 };
