+++ dev/ag-grid-angular/src/tests/utils/helpers.ts
@@ -1,4 +1,4 @@
-// eslint-disable-next-line eslint-comments/disable-enable-pair
+// eslint-disable-next-line @eslint-community/eslint-comments/disable-enable-pair
 /* eslint-disable @typescript-eslint/naming-convention */
 
 import { expect, Locator, Page } from '@playwright/test';
