+++ package.json
@@ -50,7 +50,7 @@
         "-----LINTERS-----": "----------------------------------------------------------------------------------------",
         "prettier": "prettier --check . --ignore-unknown",
         "prettier:fix": "yarn run prettier --write",
-        "eslint": "eslint . --ext='.js,.ts,.html'",
+        "eslint": "eslint .",
         "eslint:fix": "yarn run eslint --fix",
         "stylelint": "stylelint '**/*.{css,scss}'",
         "stylelint:fix": "yarn run stylelint --fix",
@@ -90,14 +90,13 @@
         "@angular-devkit/build-angular": "^18.2.21",
         "@angular-devkit/core": "^18.2.18",
         "@angular-devkit/schematics": "^18.2.21",
-        "@angular-eslint/eslint-plugin": "^18.2.0",
-        "@angular-eslint/eslint-plugin-template": "^18.2.0",
-        "@angular-eslint/template-parser": "^18.2.0",
         "@angular/cli": "^18.2.21",
         "@angular/compiler-cli": "^18.2.14",
         "@angular/language-service": "^18.2.14",
         "@commitlint/cli": "^19.8.1",
         "@commitlint/config-conventional": "^19.8.1",
+        "@eslint-community/eslint-plugin-eslint-comments": "^4.7.2",
+        "@eslint/js": "^9.39.5",
         "@jest/globals": "^29.7.0",
         "@nx/angular": "^20.8.4",
         "@nx/jest": "^20.8.4",
@@ -108,22 +107,19 @@
         "@swc/core": "^1.16.1",
         "@testing-library/angular": "16.0.0",
         "@testing-library/jest-dom": "^6.9.1",
-        "@types/eslint": "^8.56.12",
+        "@types/eslint": "^9.6.1",
         "@types/jest": "^29.5.12",
         "@types/node": "^24.10.4",
-        "@typescript-eslint/eslint-plugin": "^8.68.0",
-        "@typescript-eslint/parser": "^8.68.0",
-        "@typescript-eslint/types": "^8.68.0",
-        "@typescript-eslint/utils": "^8.68.0",
+        "angular-eslint": "^18.4.3",
         "concurrently": "^9.2.4",
-        "eslint": "^8.57.0",
+        "eslint": "^9.39.5",
         "eslint-config-prettier": "^10.1.8",
-        "eslint-plugin-eslint-comments": "^3.2.0",
         "eslint-plugin-jest": "^29.16.5",
         "eslint-plugin-playwright": "^2.11.0",
         "eslint-plugin-prettier": "^5.5.6",
         "eslint-plugin-promise": "^7.3.0",
-        "eslint-plugin-rxjs": "^5.0.3",
+        "eslint-plugin-rxjs-x": "~0.9.6",
+        "globals": "^17.12.0",
         "husky": "^9.1.7",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
@@ -141,6 +137,7 @@
         "stylelint-prettier": "^5.0.3",
         "ts-node": "^10.9.2",
         "tslib": "^2.3.0",
-        "typescript": "5.5.4"
+        "typescript": "5.5.4",
+        "typescript-eslint": "^8.68.0"
     }
 }
